requirements.txt
aimlbotkernel
ipykernel
ipython
wikipedia
matplotlib
Pillow
nltk
pandas